duplicate behavior of imageMetadata in videoMetadata #1187
