+++ src/components/NavItems/tools/Metadata/api/imageMetadataApi.jsx
@@ -1,8 +1,7 @@
 import { isValidUrl } from "@Shared/Utils/URLUtils";
-import { parseMetadata } from "@uswriting/exiftool";
 import exifr from "exifr";
 
-import { groupExifToolOutput, mergeExifrFallback } from "./exifToolUtils";
+import { mergeExifrFallback, runExifTool } from "./exifToolUtils";
 
 const exifrOptions = {
   exif: false,
@@ -12,36 +11,6 @@ const exifrOptions = {
   tiff: true,
   mergeOutput: false,
 };
-
-// Redirects zeroperl.wasm requests to the extension's bundled copy.
-// Without this "trick" the extension is not able to access to the wasm file which
-// contains everything we need to retrieve metadata
-const extensionFetch = (input, init) => {
-  const url = typeof input === "string" ? input : (input.url ?? input);
-  if (typeof url === "string" && url.endsWith("zeroperl.wasm")) {
-    return fetch(chrome.runtime.getURL("zeroperl.wasm"), init);
-  }
-  return fetch(input, init);
-};
-
-// From a file, run exif tool command with args that allows to retrieve every metadata possible
-async function runExifTool(file) {
-  const result = await parseMetadata(file, {
-    args: ["-json", "-G", "-n", "-a", "-u"],
-    fetch: extensionFetch,
-    transform: (data) => {
-      const parsed = JSON.parse(data);
-      return Array.isArray(parsed) ? parsed[0] : parsed;
-    },
-  });
-
-  if (!result.success) {
-    console.error("ExifTool error:", result.error);
-    return null;
-  }
-
-  return groupExifToolOutput(result.data);
-}
 
 /**
  * Function that retrieve metadata from a file in input
