+++ src/components/NavItems/tools/Metadata/api/videoMetadataApi.jsx
@@ -1,40 +1,32 @@
 import { isValidUrl } from "@Shared/Utils/URLUtils";
-import { parseMetadata } from "@uswriting/exiftool";
 
-import zeroPerlWasmUrl from "/zeroperl.wasm?url";
+import { runExifTool } from "./exifToolUtils";
 
-const wasmFetch = (url) =>
-  fetch(url.endsWith("zeroperl.wasm") ? zeroPerlWasmUrl : url);
-
+/**
+ * Retrieves metadata from a video URL using ExifTool.
+ * @param {string} url
+ * @returns {Promise<Record<string, Record<string,unknown>>>}
+ */
 export const extractVideoMetadata = async (url) => {
   if (!isValidUrl(url)) {
     throw new Error("Invalid URL");
   }
 
-  const response = await fetch(url, {
-    mode: "cors",
-  });
+  const response = await fetch(url, { mode: "cors" });
 
   if (!response.ok) {
     throw new Error(`HTTP error ${response.status}`);
   }
 
   const blob = await response.blob();
+  const filename = url.split("/").pop()?.split("?")[0] || "video";
+  const file = new File([blob], filename, { type: blob.type });
 
-  const jsonMetadata = await parseMetadata(
-    new File([blob], "video-file", {
-      type: blob.type,
-    }),
-    {
-      args: ["-a", "-g1", "-json", "-n"],
-      transform: (data) => JSON.parse(data),
-      fetch: wasmFetch,
-    },
-  );
+  const result = await runExifTool(file, ["-json", "-G", "-n", "-a", "-u"]);
 
-  if (!jsonMetadata.success) {
+  if (!result) {
     throw new Error("No metadata found.");
   }
-  console.log(jsonMetadata.data[0]);
-  return jsonMetadata.data[0];
+
+  return result;
 };
